+++ client/bindings/nodejs/types/index.ts
@@ -7,6 +7,7 @@ export * from './generateAddressesOptions';
 export * from './ledgerNanoStatus';
 export * from './network';
 export * from './nodeInfo';
+export * from './outputIdsResponse';
 export * from './outputBuilderOptions';
 export * from './preparedTransactionData';
 export * from './queryParameters';