+++ client/examples/output/foundry.rs
@@ -200,7 +200,7 @@ async fn main() -> Result<()> {
     ];
 
     // get additional input for the new basic output
-    let output_ids = client
+    let output_ids_response = client
         .basic_output_ids(vec![QueryParameter::Address(
             address.to_bech32(client.get_bech32_hrp().await?),
         )])
@@ -209,7 +209,7 @@ async fn main() -> Result<()> {
     let block = client
         .block()
         .with_secret_manager(&secret_manager)
-        .with_input(output_ids[0].into())?
+        .with_input(output_ids_response.items[0].into())?
         .with_input(alias_output_id.into())?
         .with_input(foundry_output_id.into())?
         .with_outputs(outputs)?