+++ client/examples/quorum.rs
@@ -34,7 +34,7 @@ async fn main() -> Result<()> {
         .await?;
 
     // Get output ids of outputs that can be controlled by this address without further unlock constraints
-    let output_ids = client
+    let output_ids_response = client
         .basic_output_ids(vec![
             QueryParameter::Address(addresses[0].clone()),
             QueryParameter::HasExpiration(false),
@@ -42,7 +42,7 @@ async fn main() -> Result<()> {
             QueryParameter::HasStorageDepositReturn(false),
         ])
         .await?;
-    println!("Address outputs: {output_ids:?}");
+    println!("Address outputs: {output_ids_response:?}");
 
     Ok(())
 }