+++ client/src/message_interface/message_handler.rs
@@ -23,7 +23,7 @@ use iota_types::block::{
 use zeroize::Zeroize;
 #[cfg(feature = "mqtt")]
 use {
-    crate::{MqttPayload, Topic},
+    crate::mqtt::{MqttPayload, Topic},
     iota_types::block::payload::milestone::option::dto::ReceiptMilestoneOptionDto,
 };
 
@@ -226,7 +226,7 @@ impl ClientMessageHandler {
                     native_tokens,
                     &alias_id,
                     state_index,
-                    state_metadata.map(|s| prefix_hex::decode(&s)).transpose()?,
+                    state_metadata.map(prefix_hex::decode).transpose()?,
                     foundry_counter,
                     unlock_conditions,
                     features,
@@ -537,18 +537,18 @@ impl ClientMessageHandler {
             Message::GetIncludedBlockMetadata { transaction_id } => Ok(Response::BlockMetadata(
                 self.client.get_included_block_metadata(&transaction_id).await?,
             )),
-            Message::BasicOutputIds { query_parameters } => Ok(Response::OutputIds(
+            Message::BasicOutputIds { query_parameters } => Ok(Response::OutputIdsResponse(
                 self.client.basic_output_ids(query_parameters).await?,
             )),
-            Message::AliasOutputIds { query_parameters } => Ok(Response::OutputIds(
+            Message::AliasOutputIds { query_parameters } => Ok(Response::OutputIdsResponse(
                 self.client.alias_output_ids(query_parameters).await?,
             )),
             Message::AliasOutputId { alias_id } => Ok(Response::OutputId(self.client.alias_output_id(alias_id).await?)),
-            Message::NftOutputIds { query_parameters } => {
-                Ok(Response::OutputIds(self.client.nft_output_ids(query_parameters).await?))
-            }
+            Message::NftOutputIds { query_parameters } => Ok(Response::OutputIdsResponse(
+                self.client.nft_output_ids(query_parameters).await?,
+            )),
             Message::NftOutputId { nft_id } => Ok(Response::OutputId(self.client.nft_output_id(nft_id).await?)),
-            Message::FoundryOutputIds { query_parameters } => Ok(Response::OutputIds(
+            Message::FoundryOutputIds { query_parameters } => Ok(Response::OutputIdsResponse(
                 self.client.foundry_output_ids(query_parameters).await?,
             )),
             Message::FoundryOutputId { foundry_id } => {