+++ client/CHANGELOG.md
@@ -19,6 +19,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 2.0.1-rc.8 - 2023-XX-XX
+
+### Added
+
+- `QueryParameters::empty()`;
+- MQTT `Error`;
+
+### Changed
+
+- `Client::{alias_output_ids, basic_output_ids, foundry_output_ids, nft_output_ids}` will not do automatic pagination if `QueryParameter::Cursor(_)` is provided and return `OutputIdsResponse`;
+- `Message::{AliasOutputIds, BasicOutputIds, FoundryOutputIds, NftOutputIds}` return `OutputIdsResponse`;
+- `Response::OutputIds` to `Response::OutputIdsResponse`;
+- Renamed `Client::get_output_ids_with_pagination()` to `Client::get_output_ids()`;
+- All MQTT related functions return an MQTT `Error`;
+- Re-export `mqtt` module instead of all its symbols;
+
 ## 2.0.1-rc.7 - 2023-03-09
 
 ### Added