+++ types/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.0.0-rc.8 - 2023-XX-XX
+
+### Changed
+
+- `OutputIdsResponse::items` from `Vec<String>` to `Vec<OutputId>`;
+
 ## 1.0.0-rc.7 - 2023-03-09
 
 ### Changed